Refactor Gradio apps to reduce code duplication
- Created `apps/ui_utils.py` for shared utility functions and logic.
- Created `apps/ui_constants.py` for shared UI constants and default texts.
- Refactored `apps/gradio_main.py` and `apps/gradio_xpu.py` to use these shared modules.
- Unified `cleanup_gpu_memory` to handle CUDA, MPS, and XPU.
- Fixed fallback logic path in `apps/gradio_xpu.py`.
- Fixed missing import in `src/vieneu/turbo.py`.
